+++ src/main/java/group/msg/jpowermonitor/MeasureMethod.java
@@ -2,10 +2,7 @@ package group.msg.jpowermonitor;
 
 import group.msg.jpowermonitor.dto.DataPoint;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
-import java.math.BigDecimal;
-import java.nio.file.Path;
 import java.util.List;
 import java.util.Map;
 
@@ -20,7 +17,10 @@ public interface MeasureMethod {
      * @return all data point for the configured paths.
      * @throws JPowerMonitorException if measurement tool is not available.
      */
-    @NotNull List<DataPoint> measure() throws JPowerMonitorException;
+    @NotNull
+    default List<DataPoint> measure() throws JPowerMonitorException {
+        return List.of(measureFirstConfiguredPath());
+    }
 
     /**
      * Measure only the first configured path.
@@ -28,12 +28,14 @@ public interface MeasureMethod {
      * @return the first data point in the configured paths.
      * @throws JPowerMonitorException if measurement tool is not available.
      */
-    @NotNull DataPoint measureFirstConfiguredPath() throws JPowerMonitorException;
+    @NotNull
+    DataPoint measureFirstConfiguredPath() throws JPowerMonitorException;
 
     /**
      * @return list of configured sensor paths.
      */
-    @NotNull List<String> configuredSensors();
+    @NotNull
+    List<String> configuredSensors();
 
     /**
      * The map of configured sensor paths with their configured default energy.
@@ -42,40 +44,6 @@ public interface MeasureMethod {
      *
      * @return Map of paths with default energy in idle mode (from config).
      */
-    @NotNull Map<String, BigDecimal> defaultEnergyInIdleModeForMeasuredSensors();
-
-    /**
-     * @return configured sampling interval
-     */
-    int getSamplingInterval();
-
-    /**
-     * @return configured init cycles
-     */
-    int initCycles();
-
-    /**
-     * @return configured sampling interval for init cycles
-     */
-    int getSamplingIntervalForInit();
-
-    /**
-     * @return configured calm down interval in ms.
-     */
-    int getCalmDownIntervalInMs();
-
-    /**
-     * @return configured path to result csv.
-     */
-    @Nullable Path getPathToResultCsv();
-
-    /**
-     * @return configured path to measurement csv.
-     */
-    @Nullable Path getPathToMeasurementCsv();
-
-    /**
-     * @return Configured percentage of samples at the beginning of the measurement to discard.
-     */
-    @NotNull BigDecimal getPercentageOfSamplesAtBeginningToDiscard();
+    @NotNull
+    Map<String, Double> defaultEnergyInIdleModeForMeasuredSensors();
 }