+++ src/main/java/group/msg/jpowermonitor/MeasureMethodProvider.java
@@ -2,6 +2,7 @@ package group.msg.jpowermonitor;
 
 import group.msg.jpowermonitor.config.JPowerMonitorConfig;
 import group.msg.jpowermonitor.measurement.csv.CommaSeparatedValuesReader;
+import group.msg.jpowermonitor.measurement.est.EstimationReader;
 import group.msg.jpowermonitor.measurement.lhm.LibreHardwareMonitorReader;
 
 /**
@@ -15,6 +16,8 @@ public class MeasureMethodProvider {
             return new CommaSeparatedValuesReader(config);
         } else if ("lhm".equals(config.getMeasurement().getMethod())) {
             return new LibreHardwareMonitorReader(config);
+        } else if ("est".equals(config.getMeasurement().getMethod())) {
+            return new EstimationReader(config);
         } else {
             throw new JPowerMonitorException("Unknown measure method " + config.getMeasurement().getMethod());
         }