+++ src/main/java/group/msg/jpowermonitor/config/EstimationCfg.java
@@ -1,0 +1,12 @@
+package group.msg.jpowermonitor.config;
+
+import lombok.Data;
+
+/**
+ * Data class for estimation measurement config.
+ */
+@Data
+public class EstimationCfg {
+    private Double cpuMinWatts;
+    private Double cpuMaxWatts;
+}