+++ src/main/java/group/msg/jpowermonitor/junit/JPowerMonitorExtension.java
@@ -18,7 +18,8 @@ import org.junit.jupiter.api.extension.BeforeEachCallback;
 import org.junit.jupiter.api.extension.ExtensionContext;
 
 import java.lang.reflect.Field;
-import java.math.BigDecimal;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -31,7 +32,6 @@ import java.util.TimerTask;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
-import static group.msg.jpowermonitor.util.Constants.MATH_CONTEXT;
 
 /**
  * JUnit Extension for measuring energy.
@@ -44,8 +44,9 @@ public class JPowerMonitorExtension implements BeforeAllCallback, BeforeEachCall
     private Timer timer;
     private TimerTask timedMeasurement;
     private long timeBeforeTest;
+    private JPowerMonitorConfig config;
     private MeasureMethod measureMethod;
-    private Map<String, BigDecimal> energyInIdleMode;
+    private Map<String, Double> energyInIdleMode;
     private ResultsWriter resultsWriter;
 
     @Override
@@ -54,9 +55,11 @@ public class JPowerMonitorExtension implements BeforeAllCallback, BeforeEachCall
         DefaultConfigProvider.invalidateCachedConfig();
 
         String configFile = context.getTestClass().map(c -> c.getSimpleName() + ".yaml").orElse(null);
-        JPowerMonitorConfig config = new DefaultConfigProvider().readConfig(configFile);
+        config = new DefaultConfigProvider().readConfig(configFile);
         measureMethod = MeasureMethodProvider.resolveMeasureMethod(config);
-        resultsWriter = new ResultsWriter(measureMethod.getPathToResultCsv(), measureMethod.getPathToMeasurementCsv(), config.getCarbonDioxideEmissionFactor());
+        final Path pathToResultCsv = config.getCsvRecording().getResultCsv() != null ? Paths.get(config.getCsvRecording().getResultCsv()) : null;
+        final Path pathToMeasurementCsv = config.getCsvRecording().getMeasurementCsv() != null ? Paths.get(config.getCsvRecording().getMeasurementCsv()) : null;
+        resultsWriter = new ResultsWriter(pathToResultCsv, pathToMeasurementCsv, config.getCarbonDioxideEmissionFactor());
         energyInIdleMode = measureIdleMode();
     }
 
@@ -64,8 +67,8 @@ public class JPowerMonitorExtension implements BeforeAllCallback, BeforeEachCall
     public void beforeEach(ExtensionContext context) throws Exception {
         powerMeasurements.clear();
         measureMethod.configuredSensors().forEach(k -> powerMeasurements.put(k, new ArrayList<>())); // init result map with configured keys
-        System.out.println("sleeping for " + measureMethod.getCalmDownIntervalInMs() + "ms in order to calm down");
-        TimeUnit.MILLISECONDS.sleep(measureMethod.getCalmDownIntervalInMs());
+        System.out.println("sleeping for " + config.getCalmDownIntervalInMs() + "ms in order to calm down");
+        TimeUnit.MILLISECONDS.sleep(config.getCalmDownIntervalInMs());
         timer = new Timer();
         timedMeasurement = new TimerTask() {
             @Override
@@ -74,7 +77,7 @@ public class JPowerMonitorExtension implements BeforeAllCallback, BeforeEachCall
                 dataPoints.forEach(dp -> powerMeasurements.get(dp.getName()).add(dp));
             }
         };
-        timer.schedule(timedMeasurement, measureMethod.getSamplingInterval(), measureMethod.getSamplingInterval());
+        timer.schedule(timedMeasurement, config.getSamplingIntervalInMs(), config.getSamplingIntervalInMs());
         timeBeforeTest = System.nanoTime();
     }
 
@@ -90,7 +93,7 @@ public class JPowerMonitorExtension implements BeforeAllCallback, BeforeEachCall
         for (Map.Entry<String, List<DataPoint>> entry : powerMeasurements.entrySet()) {
             List<DataPoint> dataPoints = new ArrayList<>(entry.getValue()); // clone list in order to avoid ConcurrentModification
             // cut off the first x% measurements
-            List<DataPoint> dataPointsToConsider = dataPoints.subList(firstXPercent(dataPoints.size(), measureMethod.getPercentageOfSamplesAtBeginningToDiscard()), dataPoints.size());
+            List<DataPoint> dataPointsToConsider = dataPoints.subList(firstXPercent(dataPoints.size(), config.getPercentageOfSamplesAtBeginningToDiscard()), dataPoints.size());
             DataPoint average = calculateAvg(dataPointsToConsider);
             SensorValue sensorValue = calculateResult(timeTaken, energyInIdleMode.get(entry.getKey()), average);
             sensorValues.add(sensorValue);
@@ -141,14 +144,12 @@ public class JPowerMonitorExtension implements BeforeAllCallback, BeforeEachCall
         return sensorValueFields;
     }
 
-    private SensorValue calculateResult(long timeTaken, BigDecimal powerInIdleMode, DataPoint dp) {
-        BigDecimal valueWithoutIdlePower = dp.getValue().subtract(powerInIdleMode);
-        BigDecimal valuePerHour = dp.isPowerSensor() ?
-            valueWithoutIdlePower.multiply(HumanReadableTime.nanosToHours(timeTaken), MATH_CONTEXT).setScale(5, MATH_CONTEXT.getRoundingMode())
-            : BigDecimal.ZERO;
-        BigDecimal valueWithIdlePowerPerHour = dp.isPowerSensor() ?
-            dp.getValue().multiply(HumanReadableTime.nanosToHours(timeTaken), MATH_CONTEXT).setScale(5, MATH_CONTEXT.getRoundingMode())
-            : BigDecimal.ZERO;
+    private SensorValue calculateResult(long timeTaken, double powerInIdleMode, DataPoint dp) {
+        double valueWithoutIdlePower = dp.getValue() - powerInIdleMode;
+        double valuePerHour = dp.isPowerSensor() ?
+            valueWithoutIdlePower * roundScale5(HumanReadableTime.nanosToHours(timeTaken))
+            : 0.0;
+        double valueWithIdlePowerPerHour = dp.isPowerSensor() ? dp.getValue() * roundScale5(HumanReadableTime.nanosToHours(timeTaken)) : 0.0;
         return SensorValue.builder().name(dp.getName())
             .value(valueWithoutIdlePower)
             .unit(dp.getUnit())
@@ -160,8 +161,8 @@ public class JPowerMonitorExtension implements BeforeAllCallback, BeforeEachCall
             .build();
     }
 
-    private Map<String, BigDecimal> measureIdleMode() {
-        Map<String, BigDecimal> defaults = measureMethod.defaultEnergyInIdleModeForMeasuredSensors();
+    private Map<String, Double> measureIdleMode() {
+        Map<String, Double> defaults = measureMethod.defaultEnergyInIdleModeForMeasuredSensors();
         defaults.forEach((k, v) -> System.out.printf("(configured) energy consumption in idle mode for %s is %s%n", k, v));
         if (defaults.size() == measureMethod.configuredSensors().size()) {
             return defaults; // then we are done
@@ -175,13 +176,13 @@ public class JPowerMonitorExtension implements BeforeAllCallback, BeforeEachCall
         Set<String> defaultsNames = defaults.keySet();
         if (powerSensorNames.equals(defaultsNames)) {
             Set<String> nonPowerSensorNames = dataPointsSample.stream().filter(x -> !x.isPowerSensor()).map(DataPoint::getName).collect(Collectors.toSet());
-            nonPowerSensorNames.forEach(x -> defaults.putIfAbsent(x, BigDecimal.ZERO)); // add zero, if not a power sensor otherwise the value is not initialized in the map!
+            nonPowerSensorNames.forEach(x -> defaults.putIfAbsent(x, 0.0)); // add zero, if not a power sensor otherwise the value is not initialized in the map!
             return defaults; // then we are done
         }
 
         long timeBeforeTest = System.nanoTime();
-        for (int i = 0; i < measureMethod.initCycles(); i++) {
-            if (sleep(measureMethod.getSamplingIntervalForInit())) {  // first sleep, then measure.
+        for (int i = 0; i < config.getInitCycles(); i++) {
+            if (sleep(config.getSamplingIntervalForInitInMs())) {  // first sleep, then measure.
                 break;
             }
             List<DataPoint> dataPoints = measureMethod.measure();
@@ -192,13 +193,13 @@ public class JPowerMonitorExtension implements BeforeAllCallback, BeforeEachCall
         return defaults;
     }
 
-    private void fillDefaultMeasurements(Map<String, BigDecimal> defaults, Map<String, List<DataPoint>> measurements) {
+    private void fillDefaultMeasurements(Map<String, Double> defaults, Map<String, List<DataPoint>> measurements) {
         for (Map.Entry<String, List<DataPoint>> entry : measurements.entrySet()) {
             List<DataPoint> dataPoints = new ArrayList<>(entry.getValue()); // clone list in order to avoid ConcurrentModification
-            List<DataPoint> dataPointsToConsider = dataPoints.subList(firstXPercent(dataPoints.size(), measureMethod.getPercentageOfSamplesAtBeginningToDiscard()), dataPoints.size());  // cut off the first x% measurements
+            List<DataPoint> dataPointsToConsider = dataPoints.subList(firstXPercent(dataPoints.size(), config.getPercentageOfSamplesAtBeginningToDiscard()), dataPoints.size());  // cut off the first x% measurements
             DataPoint average = calculateAvg(dataPointsToConsider);
             resultsWriter.writeToMeasurementCsv("Initialize", dataPointsToConsider, "(measure idle power)");
-            BigDecimal prev = defaults.putIfAbsent(entry.getKey(), average.isPowerSensor() ? average.getValue() : BigDecimal.ZERO); // add zero, if not a power sensor!
+            Double prev = defaults.putIfAbsent(entry.getKey(), average.isPowerSensor() ? average.getValue() : 0.0); // add zero, if not a power sensor!
             if (prev == null) { // then the key was not present in the map => log entry.
                 System.out.printf("(measured) %s in idle mode for %s is %s%n", average.isPowerSensor() ? "energy consumption" : "sensor value", entry.getKey(), average.getValue());
             }
@@ -217,15 +218,11 @@ public class JPowerMonitorExtension implements BeforeAllCallback, BeforeEachCall
 
     private DataPoint calculateAvg(@NotNull List<DataPoint> dataPoints) {
         if (dataPoints.isEmpty()) {
-            return new DataPoint("No Datapoints for Average", BigDecimal.ZERO, Unit.NONE, LocalDateTime.now(), null);
+            return new DataPoint("No Datapoints for Average", 0.0, Unit.NONE, LocalDateTime.now(), null);
         }
         DataPoint reference = dataPoints.get(0);
-        BigDecimal avg = dataPoints.stream()
-            .map(DataPoint::getValue)
-            .reduce(BigDecimal.ZERO, BigDecimal::add)
-            .divide(new BigDecimal(dataPoints.size()), MATH_CONTEXT)
-            .setScale(2, MATH_CONTEXT.getRoundingMode());
-        return new DataPoint(reference.getName(), avg, reference.getUnit(), LocalDateTime.now(), reference.getThreadName());
+        double avg = dataPoints.stream().mapToDouble(DataPoint::getValue).average().orElse(0.0);
+        return new DataPoint(reference.getName(), roundScale2(avg), reference.getUnit(), LocalDateTime.now(), reference.getThreadName());
     }
 
     private String getTestName(ExtensionContext context) {
@@ -241,13 +238,24 @@ public class JPowerMonitorExtension implements BeforeAllCallback, BeforeEachCall
 
     /**
      * Calculates the first x percent of an integer value.
-     * @param baseSize the base size
+     *
+     * @param baseSize   the base size
      * @param percentage the percent value to be used. For negative percentages return 0.
      * @return the first percentage % of size.
      */
-    int firstXPercent(int baseSize, BigDecimal percentage) {
-        BigDecimal positivePercentage = percentage.max(BigDecimal.ZERO);
-        BigDecimal xPercent = new BigDecimal(String.valueOf(baseSize)).multiply(positivePercentage, MATH_CONTEXT).divide(Constants.ONE_HUNDRED, MATH_CONTEXT);
-        return xPercent.setScale(0, MATH_CONTEXT.getRoundingMode()).intValue();
+    int firstXPercent(int baseSize, double percentage) {
+        double positivePercentage = Math.max(0.0, percentage);
+        double xPercent = baseSize * positivePercentage / Constants.ONE_HUNDRED;
+        return (int) xPercent;
     }
+
+    double roundScale2(double toBeRounded) {
+        return (int) (toBeRounded * 100 + 0.5) / 100.0;
+    }
+
+    double roundScale5(double toBeRounded) {
+        return (int) (toBeRounded * 100000 + 0.5) / 100000.0;
+    }
+
+
 }