+++ src/main/java/group/msg/jpowermonitor/measurement/lhm/LibreHardwareMonitorReader.java
@@ -14,12 +14,8 @@ import org.apache.hc.client5.http.classic.methods.HttpGet;
 import org.apache.hc.client5.http.impl.classic.HttpClientBuilder;
 import org.apache.hc.core5.http.ClassicHttpResponse;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 import java.io.IOException;
-import java.math.BigDecimal;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -34,12 +30,10 @@ import java.util.stream.Collectors;
  * @see MeasureMethod
  */
 public class LibreHardwareMonitorReader implements MeasureMethod {
-    HttpClient client;
-    JPowerMonitorConfig config;
-    LibreHardwareMonitorCfg lhmConfig;
+    private final HttpClient client;
+    private final LibreHardwareMonitorCfg lhmConfig;
 
     public LibreHardwareMonitorReader(JPowerMonitorConfig config) {
-        this.config = config;
         Objects.requireNonNull(config.getMeasurement().getLhm(), "Libre Hardware Monitor config must be set!");
         this.lhmConfig = config.getMeasurement().getLhm();
         this.client = HttpClientBuilder.create().build();
@@ -92,7 +86,7 @@ public class LibreHardwareMonitorReader implements MeasureMethod {
             throw new JPowerMonitorException("Unable to find element for path " + pathElement.getPath() + "!");
         }
         String[] valueAndUnit = elem.getValue().split("\\s+");// (( "5,4 W" ))
-        BigDecimal value = new BigDecimal(valueAndUnit[0].replace(',', '.').trim());
+        Double value = Double.valueOf(valueAndUnit[0].replace(',', '.').trim());
         Unit unit = Unit.fromAbbreviation(valueAndUnit[1].trim());
         return new DataPoint(String.join("->", pathElement.getPath()), value, unit, time, null);
     }
@@ -106,47 +100,12 @@ public class LibreHardwareMonitorReader implements MeasureMethod {
     }
 
     @Override
-    public @NotNull Map<String, BigDecimal> defaultEnergyInIdleModeForMeasuredSensors() {
-        Map<String, BigDecimal> energyInIdleModeForMeasuredSensors = new HashMap<>();
+    public @NotNull Map<String, Double> defaultEnergyInIdleModeForMeasuredSensors() {
+        Map<String, Double> energyInIdleModeForMeasuredSensors = new HashMap<>();
         lhmConfig.getPaths().stream()
             .filter(x -> x.getEnergyInIdleMode() != null)
             .forEach(p -> energyInIdleModeForMeasuredSensors.put(String.join("->", p.getPath()), p.getEnergyInIdleMode()));
         return energyInIdleModeForMeasuredSensors;
-    }
-
-    @Override
-    public int getSamplingInterval() {
-        return config.getSamplingIntervalInMs();
-    }
-
-    @Override
-    public int initCycles() {
-        return config.getInitCycles();
-    }
-
-    @Override
-    public int getSamplingIntervalForInit() {
-        return config.getSamplingIntervalForInitInMs();
-    }
-
-    @Override
-    public int getCalmDownIntervalInMs() {
-        return config.getCalmDownIntervalInMs();
-    }
-
-    @Override
-    public @Nullable Path getPathToResultCsv() {
-        return config.getCsvRecording().getResultCsv() != null ? Paths.get(config.getCsvRecording().getResultCsv()) : null;
-    }
-
-    @Override
-    public @Nullable Path getPathToMeasurementCsv() {
-        return config.getCsvRecording().getMeasurementCsv() != null ? Paths.get(config.getCsvRecording().getMeasurementCsv()) : null;
-    }
-
-    @Override
-    public @NotNull BigDecimal getPercentageOfSamplesAtBeginningToDiscard() {
-        return config.getPercentageOfSamplesAtBeginningToDiscard();
     }
 
     private DataElem findElement(DataElem root, Object[] path) {