+++ src/test/java/group/msg/jpowermonitor/junit/ResultsWriterTest.java
@@ -5,7 +5,6 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import java.io.IOException;
-import java.math.BigDecimal;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -21,15 +20,15 @@ class ResultsWriterTest {
 
     static Stream<Arguments> l10nTestConstructorValues() {
         return Stream.of(
-            arguments(new Locale("en", "US"),
+            arguments(new Locale.Builder().setLanguage("en").setRegion("US").build(),
                 "Time,Name,Sensor,Value,Unit,Baseload,Unit,Value+Baseload,Unit,Energy(Value),Unit,Energy(Value+Baseload),Unit,CO2 Value,Unit",
                 "Time,Name,Sensor,Value,Unit"),
-            arguments(new Locale("de", "DE"),
+            arguments(new Locale.Builder().setLanguage("de").setRegion("DE").build(),
                 "Uhrzeit;Name;Sensor;Wert;Einheit;Grundlast;Einheit;Wert+Grundlast;Einheit;Energie(Wert);Einheit;Energie(Wert+Grundlast);Einheit;CO2 Wert;Einheit",
                 "Uhrzeit;Name;Sensor;Wert;Einheit"),
-            arguments(new Locale("fr", "FR"),
-                "Heure,Nom,D\u00E9tecteur,Valeur,Unit\u00E9,Grundlast,Unit\u00E9,Valeur+charge de base,Unit\u00E9,Energie(valeur),Unit\u00E9,Energie(valeur+charge de base),Unit\u00E9,CO2 Valeur,Unit\u00E9",
-                "Heure,Nom,D\u00E9tecteur,Valeur,Unit\u00E9")
+            arguments(new Locale.Builder().setLanguage("fr").setRegion("FR").build(),
+                "Heure,Nom,Détecteur,Valeur,Unité,Grundlast,Unité,Valeur+charge de base,Unité,Energie(valeur),Unité,Energie(valeur+charge de base),Unité,CO2 Valeur,Unité",
+                "Heure,Nom,Détecteur,Valeur,Unité")
         );
     }
 
@@ -44,7 +43,7 @@ class ResultsWriterTest {
         Locale.setDefault(currentLocale);
         ResultsWriter.setLocaleDependentValues();
         // Act
-        new ResultsWriter(pathToResultCsv, pathToMeasurementCsv, new BigDecimal("485"));
+        new ResultsWriter(pathToResultCsv, pathToMeasurementCsv, 485.0);
         // Assert
         assertThat(Files.readString(pathToResultCsv, StandardCharsets.UTF_8)).isEqualTo(expContentResultCsv + NEW_LINE); // trim carriage-return
         assertThat(Files.readString(pathToMeasurementCsv, StandardCharsets.UTF_8)).isEqualTo(expContentMeasurementCsv + NEW_LINE); // trim carriage-return